feat: Fix Errors on LoginPage Button & LoginCardHome
