--- JS/main.js
@@ -1,5 +1,0 @@
-import { hasToken } from "./userloggin";
-import { validateSession } from "./userloggin";
-import { loginButton } from "./userloggin";
-import { logoutButton } from "./userloggin";
-import { logout } from "./userloggin";
