+++ JS/userloggin.js
@@ -1,79 +1,19 @@
-let hasToken = localStorage.getItem('token')
-console.log(hasToken);
+// let hasToken = localStorage.getItem('token')
+// console.log(hasToken);
 
+// const isLogged = () => {
+//     let thereToken = localStorage.getItem('token');
+//     return thereToken ? true : false;
+// }
 
-const loginButton = document.getElementById('login')
+const loginFormButton = document.getElementById('loginForm');
 
-// const validateSession = () => {
-//     let loginCard = document.getElementById('login-card');
-//     let logoutButton = document.getElementById('logout');
-//     hasToken = localStorage.getItem('token');
-//     if(!hasToken) {
-//         loginCard.classList.remove('d-none');
-//         logoutButton.classList.add('d-none');
-//     // } else if (hasToken) {
-//     } else {
-//         loginCard.classList.add('d-none');
-//         logoutButton.classList.remove('d-none');
-//     }
-// };
-
-// const validateSession = () => {
-//     let logoutButton = document.getElementById('logout')
-//     hasToken = localStorage.getItem('token');
-//     if(hasToken) {
-//         //loginCard.classList.remove('d-none');
-//         logoutButton.classList.remove('d-none');
-//     // } else if (hasToken) {
-//     } else {
-//         logoutButton.classList.add('d-none');
-//     }
-// };
-
-
-const validateSession = () => {
-    let logoutButton = document.getElementById('logout');
-    let hasToken = localStorage.getItem('token');
-
-    if (hasToken) {
-        logoutButton.classList.remove('d-none'); 
-    } else {
-        logoutButton.classList.add('d-none');
-    }
-};
-validateSession();
-
-
-// Agregar un evento de clic al botón de inicio de sesión
-loginButton.addEventListener('click', () => {
-    validateSession(); // Validar la sesión al hacer clic en el botón de inicio de sesión
-});
-
-// Llamada inicial para validar la sesión al cargar la página
-validateSession();
-loginButton.addEventListener('click', () => {
-    let token = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c'
+loginFormButton.addEventListener('click', () => {
+    let token = 'eyJhbGciOiJIUzI1NiIs5c'
     localStorage.setItem('token', token);
-    window.location.href = 'index.html';
-    validateSession();
+    window.location.href = '../index.html';
+    //validateSession();
 });
 
-// funcion para remover token, simular cierre de sesion
-const logout = () => {
-    // Eliminar el token del localStorage
-    localStorage.removeItem('token');
 
-    // Actualizar hasToken después de eliminar el token
-    hasToken = localStorage.getItem('token');
-
-    // Validar la sesión nuevamente después de eliminar el token
-    validateSession();
-};
-
-// logout 
-const logoutButton = document.querySelector('.btn-logout');
-if (logoutButton) {
-    logoutButton.addEventListener('click', () => {
-        logout();
-    });
-};
+// export { hasToken, validateSession, loginButton, logoutButton, logout };
