+++ VIEWS/loginPage.html
@@ -10,7 +10,6 @@
 </head>
 
 <body>
-  <button id="logout" class="btn-logout">Cerrar sesión</button>
   <div class="container mt-2 h-100">
     <div class="row w-50 mx-auto gap-3">
       <div class="d-flex flex-column text-center">
@@ -65,34 +64,33 @@
         <a href="" class="text-decoration-none">Forgot Password?</a>
       </div>
       <div>
-        <a href="index.html">
-          <button id="login" class="btn-login p-2 border rounded text-white w-100 bg-primary">
-            <span class="fw-semibold text-decoration-none text-white">Log in</span>
-        </a>
+        <button id="loginForm" class="btn-login p-2 border rounded text-white w-100 bg-primary">
+          <span class="fw-semibold text-decoration-none text-white">Log in</span>
         </button>
+        <div class="">
+          <p class="px-5 text-center text-decoration-none fst-italic">
+            By signing in, you are agreeing to our
+            <span>
+              <a class="text-decoration-none" href="">
+                privacy policy, terms of use
+              </a>
+            </span>
+            and
+            <span>
+              <a class="text-decoration-none" href=""> code of conduct. </a>
+            </span>
+          </p>
+        </div>
+        <hr />
+        <p class="text-center">New to DEV Community? Create account.</p>
       </div>
-      <div class="">
-        <p class="px-5 text-center text-decoration-none fst-italic">
-          By signing in, you are agreeing to our
-          <span>
-            <a class="text-decoration-none" href="">
-              privacy policy, terms of use
-            </a>
-          </span>
-          and
-          <span>
-            <a class="text-decoration-none" href=""> code of conduct. </a>
-          </span>
-        </p>
-      </div>
-      <hr />
-      <p class="text-center">New to DEV Community? Create account.</p>
+
     </div>
-  </div>
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
-    crossorigin="anonymous"></script>
-  <script src="JS/userloggin.js" type="module"></script>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+      crossorigin="anonymous"></script>
+    <script src="../JS/userloggin.js"></script>
 
 </body>
