+++ index.html
@@ -25,19 +25,20 @@
       <div class="d-flex justify-content-end">
         <!-- DIV .LOGOUT-BUTTON-BAR -->
         <div class="logut-button-bar me-4">
-          <button type="button" class="btn btn-outline-warning" id="logoutButton" onclick="window.open('index.html')">
+          <button type="button" class="btn btn-outline-warning d-none" id="logoutHome"
+            onclick="window.open('index.html')">
             Log out
           </button>
         </div>
         <!-- DIV .LOGIN-BUTTON-BAR -->
         <div class="login-button-bar">
-          <button type="button" class="login-button btn btn-outline-primary" id="login"
-            onclick="window.open('./VIEWS/loginPage.html')">
+          <button type="button" class="login-button btn btn-outline-primary d-none" id="loginBtnNav"
+            onclick="window.location.href = 'VIEWS/loginpage.html'">
             Log in
           </button>
           <img src="./IMG/icons8-bell-24.png" alt="Bell" />
-          <img class="img-thumbnail rounded-circle" width="35" src="./IMG/imageslocomotionLogo123445666556756.png"
-            alt="locomotion-logo" />
+          <img id="avatar-login" class="img-thumbnail rounded-circle" width="35"
+            src="./IMG/imageslocomotionLogo123445666556756.png" alt="locomotion-logo" />
         </div>
       </div>
     </div>
@@ -52,7 +53,7 @@
 
         <!-- CODIGO DE JAIRO BERMEJO-->
         <!-- card create account & login (left)-->
-        <div class='loggin-card card border-light-subtle text-start shadow-sm mx-3 col-md-auto'>
+        <div id="login-card" class='d-none loggin-card card border-light-subtle text-start shadow-sm mx-3 col-md-auto'>
           <div class='card-body'>
             <h5 class='card-title pb-3 fw-bold fs-4'>DEV Community is a community of 1,666,666 amazing
               developers
@@ -62,8 +63,8 @@
             </p>
             <div class='d-flex flex-column gap-1'>
               <button type='button' class='btn btn-outline-primary'>Create account</button>
-              <a href='./login.html' class='w-100'>
-                <button id="login" type='button' class='btn btn-login w-100'>Log in</button>
+              <a href='VIEWS/loginPage.html' class='w-100'>
+                <button id="loginhome" type='button' class='btn btn-login w-100'>Log in</button>
               </a>
             </div>
           </div>
@@ -333,7 +334,7 @@
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
     integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
     crossorigin="anonymous"></script>
-  <script src="JS/userloggin.js" type="module"></script>
+  <script src="JS/loggin.js"></script>
 </body>
 
 </html>
