+++ logintest.html
@@ -36,7 +36,7 @@
             </div>
         </div>
     </div>
-    <script src="JS/userloggin.js" type="module"></script>
+    <script src="JS/loggin.js" type="module"></script>
 </body>
 
 </html>
